tests/wpt/tests/html/browsers/browsing-the-web/history-traversal/pageswap/pageswap-push-with-redirect.html

<!DOCTYPE html>
<title>pageswap navigationactivation for push navigations with a same-origin redirect</title>
<link rel="help" href="https://html.spec.whatwg.org/">
<link rel="author" href="mailto:khushalsagar@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
const expectedUrl = location.href + '?new';

const params = new URLSearchParams(location.search);
// The page the popup navigates to.
const is_new_page = params.has('new');
// The initial page in the popup.
const is_popup_page = params.has('popup');
// The test page itself.
const is_test_page = !is_popup_page && !is_new_page;

const channel = new BroadcastChannel("testchannel");

if (is_test_page) {
  const expectedUrl = location.href + "?new";
  const expectedEvents = ["pageswap", expectedUrl, "push","from", "pagehide"];

  promise_test(async t => {
    let popup;
    onload = () => {
      window.events = [];
      popup = window.open("?popup");
    };

    await new Promise(resolve => {
      channel.addEventListener(
        "message", t.step_func(async (e) => {
          if (e.data === "nav") {
            assert_array_equals(window.events, expectedEvents, 'incorrect event order');
            popup.close();
            resolve();
          }
      }));
    });
  }, `pageswap on navigation with same-origin redirect`);
} else if (is_popup_page) {
  onload = () => {
    requestAnimationFrame(() => requestAnimationFrame(() => {
      location.href = "/common/redirect.py?location=/html/browsers/browsing-the-web/history-traversal/pageswap/pageswap-push-with-redirect.html?new";
    }));

    onpageswap = (e) => {
      window.opener.events.push("pageswap");
      if (e.viewTransition != null)
      window.opener.events.push("transition");
      window.opener.events.push(e.activation.entry.url);
      window.opener.events.push(e.activation.navigationType);
      if (e.activation.from == navigation.currentEntry)
        window.opener.events.push("from");
    };

    onpagehide = () => {
      window.opener.events.push("pagehide");
      channel.postMessage("nav");
    };
  };
}
</script>
